Fix the package scope of xfs_sr_on_hostA2 and xfs_sr_on_hostB1
It's the same old pytest bug, that requires importing package-scoped
fixtures in the package's own conftest.py for the scope to be properly
applied.

This caused the fixture to be left alive across multiple packages that
required it, closer to a session scope than to a package scope.

Package-scoped fixtures moved to pkgfixtures.py and imported in each
relevant package's conftest.py.
